Complete: Calculate and Plot the Average Sale Prices per Square Foot
Using the groupby function, filtered the SF housing data into a new DataFrame containing only the sale price per square foot and the gross rent. Then, using the DF. created a line plot, stylized and formatting in a readable and professional manner that contains a widget for easy and grouped neighborhood data vizulization.